+++ osu.Game/Overlays/BeatmapSet/Scores/TopScoreUserSection.cs
@@ -96,7 +96,7 @@ namespace osu.Game.Overlays.BeatmapSet.Scores
                             {
                                 Anchor = Anchor.CentreLeft,
                                 Origin = Anchor.CentreLeft,
-                                Font = OsuFont.GetFont(size: 10, weight: FontWeight.Bold)
+                                Font = OsuFont.GetFont(size: 14, weight: FontWeight.Bold)
                             },
                             flag = new UpdateableFlag
                             {
@@ -125,7 +125,7 @@ namespace osu.Game.Overlays.BeatmapSet.Scores
             {
                 avatar.User = value.User;
                 flag.Country = value.User.Country;
-                date.Text = $@"achieved {HumanizerUtils.Humanize(value.Date)}";
+                date.Text = $@"达成于 {HumanizerUtils.Humanize(value.Date)}";
 
                 usernameText.Clear();
                 usernameText.AddUserLink(value.User);