+++ osu.Game/Screens/Menu/IntroTrianglesCN.cs
@@ -25,7 +25,7 @@ using osuTK.Graphics;
 
 namespace osu.Game.Screens.Menu
 {
-    public class IntroTriangles : IntroScreen
+    public class IntroTrianglesCN : IntroScreen
     {
         protected override string BeatmapHash => "a1556d0801b3a6b175dda32ef546f0ec812b400499f575c44fccbe9c67f9b1e5";
 
@@ -179,17 +179,17 @@ namespace osu.Game.Screens.Menu
                 using (BeginAbsoluteSequence(0, true))
                 {
                     using (BeginDelayedSequence(text_1, true))
-                        welcomeText.FadeIn().OnComplete(t => t.Text = "wel");
+                        welcomeText.FadeIn().OnComplete(t => t.Text = "欢");
 
                     using (BeginDelayedSequence(text_2, true))
-                        welcomeText.FadeIn().OnComplete(t => t.Text = "welcome");
+                        welcomeText.FadeIn().OnComplete(t => t.Text = "欢迎");
 
                     using (BeginDelayedSequence(text_3, true))
-                        welcomeText.FadeIn().OnComplete(t => t.Text = "welcome to");
+                        welcomeText.FadeIn().OnComplete(t => t.Text = "欢迎来到");
 
                     using (BeginDelayedSequence(text_4, true))
                     {
-                        welcomeText.FadeIn().OnComplete(t => t.Text = "welcome to osu!");
+                        welcomeText.FadeIn().OnComplete(t => t.Text = "欢迎来到osu!");
                         welcomeText.TransformTo(nameof(welcomeText.Spacing), new Vector2(50, 0), 5000);
                     }