+++ osu.Game/Screens/Menu/OsuLogo.cs
@@ -42,6 +42,8 @@ namespace osu.Game.Screens.Menu
 
         private readonly IntroSequence intro;
 
+        private readonly IntroSequenceCN introCN;
+
         private SampleChannel sampleClick;
         private SampleChannel sampleBeat;
 
@@ -100,6 +102,10 @@ namespace osu.Game.Screens.Menu
             Children = new Drawable[]
             {
                 intro = new IntroSequence
+                {
+                    RelativeSizeAxes = Axes.Both,
+                },
+                introCN = new IntroSequenceCN
                 {
                     RelativeSizeAxes = Axes.Both,
                 },
@@ -317,6 +323,17 @@ namespace osu.Game.Screens.Menu
             intro.Show();
             intro.Start(length);
             intro.Delay(length + fade).FadeOut();
+        }
+
+        public void PlayIntroCN()
+        {
+            const double length = 3150;
+            const double fade = 200;
+
+            logoHoverContainer.FadeOut().Delay(length).FadeIn(fade);
+            introCN.Show();
+            introCN.Start(length);
+            introCN.Delay(length + fade).FadeOut();
         }
 
         protected override void Update()